Update build-apk.yml  Making Merge buttons match in behavior.
Merge buttons were not matchining in behavior. The merge button changes to merge trips when in merge mode.  When the merge is cancelled, it should revert back to simply Merge.  Teh categorized tab worked correctly while the classify tab did not revert back but kept merge trips on the button even if merge was cancelled.